+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Format based on [Keep a Changelog](https://keepachangelog.com/), adapted for int
 
 ## [Unreleased]
 
+## [1096] - 2026-06-04
+
+- [TASK-589] Fix: make scope add materialize sparse-checkout paths
+
 ## [1095] - 2026-06-03
 
 - [TASK-588] Hydrate auto-derived scope from referenced commit file lists